+++ emp-zk/emp-svole/mpfss_reg.h
@@ -25,9 +25,9 @@ public:
   using K = typename P::K;
 
   int party;
-  int item_n, idx_max;
-  int tree_height, leave_n;
-  int tree_n;
+  int64_t item_n, idx_max;
+  int64_t tree_height, leave_n;
+  int64_t tree_n;
   bool is_malicious;
 
   PRG prg;
@@ -38,7 +38,7 @@ public:
   std::vector<block> mac_buf;          // scratch for cggm leaves
   std::vector<uint32_t> item_pos_recver;
 
-  MpfssReg(int party, int n, int t, int log_bin_sz, IO *io) {
+  MpfssReg(int party, int64_t n, int64_t t, int64_t log_bin_sz, IO *io) {
     this->party = party;
     this->netio = io;
     this->is_malicious = false;
@@ -69,7 +69,7 @@ public:
   void mpfss(OTPre<IO> *ot, AuthValue<P> *pre, AuthValue<P> *sparse_vector) {
     std::vector<SpfssSend<IO> *> senders;
     std::vector<SpfssRecv<IO> *> recvers;
-    for (int i = 0; i < tree_n; ++i) {
+    for (int64_t i = 0; i < tree_n; ++i) {
       if (party == ALICE) {
         senders.push_back(new SpfssSend<IO>(netio, tree_height));
         ot->choices_sender();
@@ -82,7 +82,7 @@ public:
     netio->flush();
     ot->reset();
 
-    for (int i = 0; i < tree_n; ++i) {
+    for (int64_t i = 0; i < tree_n; ++i) {
       block *leaves_i = mac_buf.data() + i * leave_n;
       ggm_tree[i] = leaves_i;
       if (party == ALICE) {
@@ -109,13 +109,13 @@ public:
     }
 
     // Copy the scratch mac buffer into the AuthValue array's .mac slots.
-    for (int i = 0; i < idx_max; ++i)
+    for (int64_t i = 0; i < idx_max; ++i)
       sparse_vector[i].mac = mac_buf[i];
 
     // Receiver injects val at sparse positions.
     if (party == BOB) {
-      for (int i = 0; i < tree_n; ++i) {
-        int64_t pt = (int64_t)i * leave_n + (int64_t)item_pos_recver[i];
+      for (int64_t i = 0; i < tree_n; ++i) {
+        int64_t pt = i * leave_n + (int64_t)item_pos_recver[i];
         sparse_vector[pt].val = P::k_add(sparse_vector[pt].val, pre[i].val);
       }
     }
@@ -125,13 +125,13 @@ public:
   }
 
   // sender check
-  void consistency_batch_check(F y, int num) {
+  void consistency_batch_check(F y, int64_t num) {
     F x_star;
     netio->recv_data(&x_star, sizeof(F));
     F vb = P::f_mul(secret_share_x, x_star);
     vb = P::f_sub(vb, y);
 
-    for (int i = 0; i < num; ++i)
+    for (int64_t i = 0; i < num; ++i)
       vb = P::f_add(vb, check_VW_buf[i]);
     Hash hash;
     block h = hash.hash_for_block(&vb, sizeof(F));
@@ -140,9 +140,9 @@ public:
   }
 
   // receiver check
-  void consistency_batch_check(AuthValue<P> *pre, F z, int num) {
+  void consistency_batch_check(AuthValue<P> *pre, F z, int64_t num) {
     F beta_mul_chialpha = P::f_zero();
-    for (int i = 0; i < num; ++i) {
+    for (int64_t i = 0; i < num; ++i) {
       F tmp = P::scalar_mul(pre[i].val, check_chialpha_buf[i]);
       beta_mul_chialpha = P::f_add(beta_mul_chialpha, tmp);
     }
@@ -151,7 +151,7 @@ public:
     netio->flush();
 
     F va = z;
-    for (int i = 0; i < num; ++i)
+    for (int64_t i = 0; i < num; ++i)
       va = P::f_add(va, check_VW_buf[i]);
 
     Hash hash;
